requirements_rag.txt
gradio==5.5.0
faiss-cpu==1.9.0
numpy==1.26.4
sentence-transformers==3.0.1
beautifulsoup4==4.14.0
markdownify==0.14.1
python-dotenv==1.0.1
openai==1.52.2
google-generativeai==0.8.3
tqdm==4.66.5
scikit-learn==1.5.2
